backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
aiohttp==3.11.14
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
Deprecated==1.2.14
dnspython==2.7.0
durationpy==0.9
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
google-auth==2.35.0
greenlet==3.1.1
httpcore==1.0.9
httpx==0.25.2
h11==0.16.0
idna==3.10
importlib_metadata==6.11.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==31.0.0
limits==3.13.0
MarkupSafe==3.0.2
motor==3.6.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
passlib==1.7.4
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydantic-avro==0.7.1
PyJWT==2.9.0
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.18
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.40.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.2
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
confluent-kafka==2.6.1
avro-python3==1.10.2
fastavro==1.9.4
aiokafka==0.12.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-instrumentation-pymongo==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-propagator-b3==1.22.0
rich==13.9.4
sse-starlette==2.1.3
prometheus_client~=0.21.0
pytest~=8.3.3
prometheus_client~=0.21.0